app/backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "alembic>=1.14",
    "asyncpg>=0.30",
    "fastapi[standard]>=0.128.6",
    "greenlet>=3.3.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.0",
    "python-jose[cryptography]>=3.5.0",
    "sqlmodel>=0.0.22",
]

[dependency-groups]
dev = [
    "bandit>=1.7",
    "mypy>=1.11",
    "ruff>=0.9",
]
test = [
    "aiosqlite>=0.20",
    "httpx>=0.28",
    "pytest>=8.1",
    "pytest-asyncio>=1.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]

[[tool.mypy.overrides]]
module = "jose.*"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module = "app.services.*"
disable_error_code = ["call-overload", "arg-type", "attr-defined", "union-attr", "no-any-return", "type-arg"]

[tool.bandit]
exclude_dirs = ["tests"]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = [
    "F", "E", "W", "I", "UP", "B", "C4",
    "SIM", "FA", "ISC", "ICN", "RET", "TC", "PTH", "RUF",
]
ignore = ["ISC001", "B008"]

[tool.ruff.lint.isort]
known-first-party = ["app"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true
